Bug squashing around friend requests and related

+++ app/models/friend_request.rb
@@ -7,7 +7,10 @@ class FriendRequest < ActiveRecord::Base
 
   validates :sender_id, :recipient_id, :presence => true
   validates :sender_id, :uniqueness => { :scope => :recipient_id }
-  validate :no_two_way_requests, :recipient_exists
+  validate :no_two_way_requests, 
+           :recipient_exists, 
+           :cant_request_self,
+           :friendship_cant_exist
 
   def self.pending
     where(:pending => true)
@@ -15,7 +18,7 @@ class FriendRequest < ActiveRecord::Base
 
   def accept
     friendship = Friendship.create_friendship(recipient_id, sender_id)
-    self.destroy
+    self.destroy if friendship.persisted?
     friendship
   end
 
@@ -37,6 +40,18 @@ class FriendRequest < ActiveRecord::Base
   def recipient_exists
     unless User.exists?(self.recipient_id)
       errors[:recipient_id] << "does not exist"
+    end
+  end
+
+  def cant_request_self
+    if recipient_id == sender_id
+      errors[:sender_id] << "can't request friendship with self"
+    end
+  end
+
+  def friendship_cant_exist
+    if Friendship.exists?(:user_id => sender_id, :friend_id => recipient_id)
+      errors[:sender_id] << "can't request friendship from a friend"
     end
   end
 end
